requirements.txt
aiogram==3.4.1
python-dotenv==1.0.0
google-generativeai==0.8.6
aiohttp==3.11.14
structlog==25.5.0
diskcache==5.6.3
